I think I know what the 95% solution is finally; working toward it now in the new "PostingIndex" class which will subsume a lot of others.
There will be a "TmpPostingIndex" format which will be faster to
write and merge, and any more complicated format need only consume a
List<TmpPostingIndexReader>

+++ coders/src/main/java/edu/umass/cs/ciir/waltz/coders/map/impl/WaltzDiskMapWriter.java
@@ -21,26 +21,44 @@ import static edu.umass.cs.ciir.waltz.coders.map.impl.WaltzDiskMap.getValueFileN
  */
 public class WaltzDiskMapWriter<K, V> implements IOMapWriter<K, V> {
   final Directory outputDir;
-  final Directory sortDir;
   final FileSink valuesFile;
+
+  final boolean sorting;
+  final Directory sortDir;
   final ExternalSortingWriter<VocabEntry<K>> keySorter;
+
   final FileSink keysFile;
   final Coder<K> keyCoder;
   final Coder<V> valCoder;
   final VocabEntryCoder<K> koffCoder;
   long keyCount;
 
-  public WaltzDiskMapWriter(Directory outputDir, String baseName, Coder<K> keyCoder, Coder<V> valCoder) throws IOException {
+  public WaltzDiskMapWriter(Directory outputDir, String baseName, Coder<K> keyCoder, Coder<V> valCoder, boolean sorting) throws IOException {
     this.outputDir = outputDir;
     this.valuesFile = new FileSink(getValueFileName(outputDir, baseName));
     valuesFile.write(WaltzDiskMap.MagicCoder, "waltz.values");
     this.keysFile = new FileSink(getKeyFileName(outputDir, baseName));
     this.koffCoder = new VocabEntryCoder<>(keyCoder);
-    this.sortDir = new Directory(getKeySortDirectory(outputDir, baseName));
-    this.keySorter = new ExternalSortingWriter<>(sortDir.get(), koffCoder);
     this.keyCoder = keyCoder.lengthSafe();
     this.valCoder = valCoder; // valCoder need not be length-safe.
+    this.sorting = sorting;
     keyCount = 0;
+
+    if(sorting) {
+      this.sortDir = new Directory(getKeySortDirectory(outputDir, baseName));
+      this.keySorter = new ExternalSortingWriter<>(sortDir.get(), koffCoder);
+    } else {
+      this.sortDir = null;
+      this.keySorter = null;
+
+      // need to patch-up the keyCount at the end!
+      keysFile.write(WaltzDiskMap.MagicCoder, "waltz.keys.naive");
+      keysFile.write(FixedSize.longs, keyCount);
+    }
+  }
+
+  public WaltzDiskMapWriter(Directory outputDir, String baseName, Coder<K> keyCoder, Coder<V> valCoder) throws IOException {
+    this(outputDir, baseName, keyCoder, valCoder, true);
   }
 
   @Override
@@ -58,17 +76,17 @@ public class WaltzDiskMapWriter<K, V> implements IOMapWriter<K, V> {
    */
   public void beginWrite(K key) throws IOException {
     long startVal = valuesFile.tell();
-    keySorter.process(new VocabEntry<>(key, startVal));
+    if(sorting) {
+      keySorter.process(new VocabEntry<>(key, startVal));
+    }
     keyCount++;
   }
 
   @Override
   public void putUnsafe(K key, DataChunk val) throws IOException {
-    long startVal = valuesFile.tell();
+    beginWrite(key);
     valuesFile.write(val);
     IO.close(val);
-    keySorter.process(new VocabEntry<>(key, startVal));
-    keyCount++;
   }
 
   @Override
@@ -79,18 +97,24 @@ public class WaltzDiskMapWriter<K, V> implements IOMapWriter<K, V> {
   @Override
   public void close() throws IOException {
     this.valuesFile.close();
-    this.keySorter.close();
+    if(sorting) {
+      this.keySorter.close();
 
-    // TODO, replace this with a class/interface so you can store keys in many ways, and maybe a smart factory that chooses efficient implementations based on type.
-    keysFile.write(WaltzDiskMap.MagicCoder, "waltz.keys.naive");
-    keysFile.write(FixedSize.longs, keyCount);
-    for (VocabEntry<K> kv : this.keySorter.getOutput()) {
-      keysFile.write(koffCoder, kv);
+      // TODO, replace this with a class/interface so you can store keys in many ways, and maybe a smart factory that chooses efficient implementations based on type.
+      keysFile.write(WaltzDiskMap.MagicCoder, "waltz.keys.naive");
+      keysFile.write(FixedSize.longs, keyCount);
+      for (VocabEntry<K> kv : this.keySorter.getOutput()) {
+        keysFile.write(koffCoder, kv);
+      }
+    } else {
+      // Fix up the keyCount in the header:
+      keysFile.seekAbsolute(WaltzDiskMap.MagicLength);
+      keysFile.write(FixedSize.longs, keyCount);
+
+      // clean up sort files:
+      sortDir.removeRecursively();
     }
     keysFile.close();
-
-    // clean up sort files:
-    sortDir.removeRecursively();
   }
 
   @Override
@@ -105,7 +129,9 @@ public class WaltzDiskMapWriter<K, V> implements IOMapWriter<K, V> {
 
   @Override
   public void flush() throws IOException {
-    keySorter.flush();
+    if(sorting) {
+      keySorter.flush();
+    }
     keysFile.flush();
     valuesFile.flush();
   }
